+++ src/main/java/net/kyori/registry/DefaultedRegistryImpl.java
@@ -21,23 +21,32 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.registry.impl.registry;
+package net.kyori.registry;
 
-import net.kyori.registry.ForwardingRegistry;
-import net.kyori.registry.Registry;
-import net.kyori.registry.map.IncrementalIdMap;
+import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.checker.nullness.qual.NonNull;
-import org.checkerframework.checker.nullness.qual.Nullable;
 
-import java.util.OptionalInt;
-
-public class IdRegistryImpl<K, V> implements ForwardingRegistry<K, V> {
+/**
+ * A {@link DefaultedRegistryImpl} is an extension of a {@link RegistryImpl}. This type of registry provides
+ * a default values, as registered, for missing keys.
+ *
+ * @param <K> the key type
+ * @param <V> the value type
+ */
+public class DefaultedRegistryImpl<K, V> implements DefaultedRegistry<K, V>, ForwardingRegistry<K, V> {
   private final Registry<K, V> registry;
-  protected final IncrementalIdMap<V> ids;
+  private final K defaultKey;
+  private @MonotonicNonNull V defaultValue;
 
-  public IdRegistryImpl(final @NonNull Registry<K, V> registry, final @NonNull IncrementalIdMap<V> ids) {
+  protected DefaultedRegistryImpl(final @NonNull Registry<K, V> registry, final @NonNull K defaultKey) {
     this.registry = registry;
-    this.ids = ids;
+    this.defaultKey = defaultKey;
+
+    registry.addRegistrationListener((key, value) -> {
+      if(defaultKey.equals(key)) {
+        this.defaultValue = value;
+      }
+    });
   }
 
   @Override
@@ -46,36 +55,13 @@ public class IdRegistryImpl<K, V> implements ForwardingRegistry<K, V> {
   }
 
   @Override
-  public @NonNull V register(final @NonNull K key, final @NonNull V value) {
-    return this.register(this.ids.next(), key, value);
+  public @NonNull K defaultKey() {
+    return this.defaultKey;
   }
 
-  // TODO: should be an interface method that we implement
-  public @NonNull V register(final int id, final @NonNull K key, final @NonNull V value) {
-    this.ids.put(id, value);
-    this.registry.register(key, value);
-    return value;
-  }
-
-  // TODO: should be an interface method that we implement
-  /**
-   * Gets the id for {@code value}.
-   *
-   * @param value the value
-   * @return the id
-   */
-  public @NonNull OptionalInt id(final @NonNull V value) {
-    return this.ids.id(value);
-  }
-
-  // TODO: should be an interface method that we implement
-  /**
-   * Gets the value for {@code id}.
-   *
-   * @param id the id
-   * @return the value
-   */
-  public @Nullable V byId(final int id) {
-    return this.ids.get(id);
+  @Override
+  public @NonNull V get(final @NonNull K key) {
+    final V value = this.registry.get(key);
+    return value != null ? value : this.defaultValue;
   }
 }
