+++ src/main/java/net/kyori/registry/map/AbstractIdMap.java
@@ -21,56 +21,40 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package net.kyori.registry.impl.registry;
+package net.kyori.registry.map;
 
-import net.kyori.registry.DefaultedRegistryGetter;
-import net.kyori.registry.ForwardingRegistry;
-import net.kyori.registry.Registry;
-import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
+import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
+import it.unimi.dsi.fastutil.objects.Object2IntMap;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 /**
- * A {@link DefaultedRegistryImpl} is an extension of a {@link RegistryImpl}. This type of registry provides
- * a default values, as registered, for missing keys.
+ * An abstract implementation of an id map.
  *
- * @param <K> the key type
  * @param <V> the value type
  */
-public class DefaultedRegistryImpl<K, V> implements DefaultedRegistryGetter<K, V>, ForwardingRegistry<K, V> {
-  private final Registry<K, V> registry;
-  private final K defaultKey;
-  private @MonotonicNonNull V defaultValue;
+public abstract class AbstractIdMap<V> implements IdMap<V> {
+  protected final Int2ObjectMap<V> idToV;
+  protected final Object2IntMap<V> vToId;
 
-  public DefaultedRegistryImpl(final @NonNull Registry<K, V> registry, final @NonNull K defaultKey) {
-    this.registry = registry;
-    this.defaultKey = defaultKey;
-
-    registry.addRegistrationListener((key, value) -> {
-      if(defaultKey.equals(key)) {
-        this.defaultValue = value;
-      }
-    });
+  protected AbstractIdMap(final @NonNull Int2ObjectMap<V> idToV, final @NonNull Object2IntMap<V> vToId) {
+    this.idToV = idToV;
+    this.vToId = vToId;
   }
 
   @Override
-  public @NonNull Registry<K, V> registry() {
-    return this.registry;
+  public final @NonNull V put(final int id, final @NonNull V value) {
+    this.put0(id, value);
+    return value;
+  }
+
+  protected void put0(final int id, final @NonNull V value) {
+    this.idToV.put(id, value);
+    this.vToId.put(value, id);
   }
 
   @Override
-  public @Nullable V get(final @NonNull K key) {
-    return this.getOrDefault(key);
-  }
-
-  @Override
-  public @NonNull K defaultKey() {
-    return this.defaultKey;
-  }
-
-  @Override
-  public @NonNull V getOrDefault(final @NonNull K key) {
-    final V value = this.registry.get(key);
-    return value != null ? value : this.defaultValue;
+  public @Nullable V get(final int id) {
+    return this.idToV.get(id);
   }
 }
